Extract NodeStatus enum out and return/pass NodeStatus directly instead of the byte value
Also do some code style clean up
